Half implemented camera... last commit before SceneManager merge

+++ src/game/SceneManager.cpp
@@ -4,6 +4,6 @@
 
 #include "SceneManager.h"
 
-void SceneManager::Update(unsigned long dt) {
+void SceneManager::Update(double dt) {
     SceneStack.top().Update(dt);
 }
